# docs/swagger.json
{
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "API сервиса сокращения ссылок",
    "title": "URL Shortener API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "API Support",
      "url": "http://www.swagger.io/support",
      "email": "support@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
    "/": {
      "post": {
        "description": "Создание короткой ссылки из тела запроса",
        "consumes": [
          "text/plain"
        ],
        "produces": [
          "text/plain"
        ],
        "tags": [
          "shorter"
        ],
        "summary": "Shorten URL (text/plain)",
        "parameters": [
          {
            "description": "URL для сокращения",
            "name": "url",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Сокращённый URL",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Некорректный запрос",
            "schema": {
              "type": "string"
            }
          },
          "409": {
            "description": "URL уже существует",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/shorten": {
      "post": {
        "description": "Создание короткой ссылки (application/json)",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "shorter"
        ],
        "summary": "Shorten URL (JSON)",
        "parameters": [
          {
            "description": "URL для сокращения",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/github_com_IgorRAzumov_go-link-shorter_internal_controller_rest_handler_shorter_model.ShortenRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/github_com_IgorRAzumov_go-link-shorter_internal_controller_rest_handler_shorter_model.ShortenResponse"
            }
          },
          "400": {
            "description": "Некорректный запрос",
            "schema": {
              "type": "string"
            }
          },
          "409": {
            "description": "URL уже существует",
            "schema": {
              "$ref": "#/definitions/github_com_IgorRAzumov_go-link-shorter_internal_controller_rest_handler_shorter_model.ShortenResponse"
            }
          }
        }
      }
    },
    "/api/shorten/batch": {
      "post": {
        "description": "Пакетное сокращение нескольких URL",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "shorter"
        ],
        "summary": "Batch shorten URLs",
        "parameters": [
          {
            "description": "Список URL",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/github_com_IgorRAzumov_go-link-shorter_internal_controller_rest_handler_shorter_model.BatchShortenRequest"
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/github_com_IgorRAzumov_go-link-shorter_internal_controller_rest_handler_shorter_model.BatchShortenResponse"
              }
            }
          },
          "400": {
            "description": "Некорректный запрос",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/user/urls": {
      "get": {
        "description": "Список сокращённых ссылок текущего пользователя",
        "produces": [
          "application/json"
        ],
        "tags": [
          "user"
        ],
        "summary": "Get user URLs",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/internal_controller_rest_handler_shorter.UserURLResponse"
              }
            }
          },
          "204": {
            "description": "Нет ссылок",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Не авторизован",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "delete": {
        "description": "Мягкое удаление ссылок по short_key",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "user"
        ],
        "summary": "Delete user URLs",
        "parameters": [
          {
            "description": "Список short_key для удаления",
            "name": "short_keys",
            "in": "body",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Принято",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Не авторизован",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/ping": {
      "get": {
        "description": "Проверка доступности базы данных",
        "tags": [
          "health"
        ],
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/{shortKey}": {
      "get": {
        "description": "Редирект 307 на полный URL по короткому ключу",
        "tags": [
          "resolver"
        ],
        "summary": "Resolve short URL",
        "parameters": [
          {
            "type": "string",
            "description": "Короткий ключ",
            "name": "shortKey",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "307": {
            "description": "Location: полный URL",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Не найден",
            "schema": {
              "type": "string"
            }
          },
          "410": {
            "description": "URL удалён",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "github_com_IgorRAzumov_go-link-shorter_internal_controller_rest_handler_shorter_model.BatchShortenRequest": {
      "type": "object",
      "properties": {
        "correlation_id": {
          "description": "ID для сопоставления с ответом",
          "type": "string"
        },
        "original_url": {
          "description": "Исходный URL",
          "type": "string"
        }
      }
    },
    "github_com_IgorRAzumov_go-link-shorter_internal_controller_rest_handler_shorter_model.BatchShortenResponse": {
      "type": "object",
      "properties": {
        "correlation_id": {
          "description": "ID для сопоставления с запросом",
          "type": "string"
        },
        "short_url": {
          "description": "Сокращённый URL",
          "type": "string"
        }
      }
    },
    "github_com_IgorRAzumov_go-link-shorter_internal_controller_rest_handler_shorter_model.ShortenRequest": {
      "type": "object",
      "properties": {
        "url": {
          "description": "Исходный URL для сокращения",
          "type": "string"
        }
      }
    },
    "github_com_IgorRAzumov_go-link-shorter_internal_controller_rest_handler_shorter_model.ShortenResponse": {
      "type": "object",
      "properties": {
        "result": {
          "description": "Сокращённый URL",
          "type": "string"
        }
      }
    },
    "internal_controller_rest_handler_shorter.UserURLResponse": {
      "type": "object",
      "properties": {
        "original_url": {
          "description": "Исходный URL",
          "type": "string"
        },
        "short_url": {
          "description": "Сокращённый URL",
          "type": "string"
        }
      }
    }
  }
}